+++ experimental/packages/opentelemetry-instrumentation-http/src/http.ts
@@ -70,6 +70,7 @@ import {
   setSpanWithError,
 } from './utils';
 import type { Err, Func, Http, HttpRequestArgs, Https } from './internal-types';
+import { DEFAULT_QUERY_STRINGS_TO_REDACT } from './internal-types';
 
 /**
  * `node:http` and `node:https` instrumentation for OpenTelemetry
@@ -550,6 +551,9 @@ export class HttpInstrumentation extends InstrumentationBase<HttpInstrumentation
           ),
           enableSyntheticSourceDetection:
             instrumentation.getConfig().enableSyntheticSourceDetection || false,
+          redactedQueryParams:
+            instrumentation.getConfig().redactedQueryParamsServer ??
+            Array.from(DEFAULT_QUERY_STRINGS_TO_REDACT),
         },
         instrumentation._diag
       );
@@ -698,8 +702,7 @@ export class HttpInstrumentation extends InstrumentationBase<HttpInstrumentation
             optionsParsed,
             instrumentation.getConfig().startOutgoingSpanHook
           ),
-          redactedQueryParams:
-            instrumentation.getConfig().redactedQueryParams, // Added config for adding custom query strings
+          redactedQueryParams: instrumentation.getConfig().redactedQueryParams, // Added config for adding custom query strings
         },
         instrumentation.getConfig().enableSyntheticSourceDetection || false
       );